+++ spring-integration-core/src/test/java/org/springframework/integration/config/annotation/SplitterAnnotationPostProcessorTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2022 the original author or authors.
+ * Copyright 2002-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,7 +20,6 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.context.Lifecycle;
 import org.springframework.integration.annotation.MessageEndpoint;
 import org.springframework.integration.annotation.Splitter;
@@ -50,6 +49,7 @@ public class SplitterAnnotationPostProcessorTests {
 
 	@BeforeEach
 	public void init() {
+		this.context.registerBean(MessagingAnnotationPostProcessor.class);
 		this.context.registerChannel("input", this.inputChannel);
 		this.context.registerChannel("output", this.outputChannel);
 	}
@@ -61,12 +61,8 @@ public class SplitterAnnotationPostProcessorTests {
 
 	@Test
 	public void testSplitterAnnotation() {
-		MessagingAnnotationPostProcessor postProcessor = new MessagingAnnotationPostProcessor();
-		postProcessor.postProcessBeanDefinitionRegistry((BeanDefinitionRegistry) this.context.getBeanFactory());
-		postProcessor.postProcessBeanFactory(this.context.getBeanFactory());
-		postProcessor.afterSingletonsInstantiated();
 		TestSplitter splitter = new TestSplitter();
-		postProcessor.postProcessAfterInitialization(splitter, "testSplitter");
+		context.registerEndpoint("testSplitter", splitter);
 		context.refresh();
 		inputChannel.send(new GenericMessage<>("this.is.a.test"));
 		Message<?> message1 = outputChannel.receive(500);
